docker/pipelines/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
scipy==1.13.0
joblib==1.4.2

SQLAlchemy==2.0.29
psycopg2-binary==2.9.9

pyyaml==6.0.2
python-dotenv==1.0.1

# parquet support for pandas.read_parquet / to_parquet
pyarrow==16.1.0
